[refactor] Engineering review improvements: RwLock, shared config, graceful shutdown, tests
- Mutex→RwLock for ENGINE singleton (reduce contention at 90Hz)
- UDP send target from TCP peer address instead of hardcoded 0.0.0.0
- Shared config via fvp_get_config() FFI (C++ reads from Rust config)
- CancellationToken for graceful async task shutdown
- RwLock/Mutex poisoning error logging
- RTP sequence number managed by RtpPacketizer instance (remove static AtomicU16)
- Add 14 new tests (FFI boundary + engine core)
- Fix u16 overflow bug in test pattern generator

+++ rust/streaming-engine/src/video/test_pattern.rs
@@ -11,7 +11,7 @@ pub fn generate_nv12_frame(width: u32, height: u32, frame_num: u64) -> Vec<u8> {
     let shift = (frame_num % 256) as u8;
     for y in 0..h {
         for x in 0..w {
-            let luma = ((x as u16 * 255 / w.max(1) as u16) as u8).wrapping_add(shift);
+            let luma = ((x * 255 / w.max(1)) as u8).wrapping_add(shift);
             buf[y * w + x] = luma;
         }
     }
